feat: style mobile navbar using common layout
